+++ NEWS.md
@@ -5,6 +5,16 @@ All notable Changes to the Julia package `LieGroups.jl` will be documented in th
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.2] unreleased
+
+### Added
+
+* `is_flat` for `SpecialEuclideanGroup`
+
+### Fixed
+
+* `get_vector` on `SpecialEuclideanGroup` with `ArrayPartition` point type.
+
 ## [0.1.1] 2025-05-05
 
 ### Added